// src/types/ActionTypes.ts
export enum ActionTypes {
  SetHotPrices = 'setHotPrices',
  SetProducts = 'setProducts',
  SetPhones = 'setPhones',
  SetTablets = 'setTablets',
  SetAccessories = 'setAccessories',
  AddToFavourites = 'addToFavourites',
  RemoveFromfavourite = 'removeFromFavourite',
  AddToCart = 'addToCart',
  RemoveFromCart = 'removeFromCart',
  PlusOneItem = 'plusOneItem',
  MinusOneItem = 'minusOneItem',
  AddSelectedProduct = 'addSelectedProduct',
  SetIsLoading = 'setIsLoading',
  ClearCart = 'clearCart',
  SetCart = 'setCart',
  SetFavourites = 'setFavourites',
  FindProduct = 'findProduct',
  SetBrandNewModels = 'setBrandNewModels',
}